+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Parser support for `goto` statements.
+- Parser support for `continue` statements.
+- Parser support for `break` statements.
+- Parser support for `for` loops.
 - Parser support for negative integer literals.
 
 ## [0.1.1] - 2021-05-21